Adding status alert

+++ database/migrations/2021_12_05_194750_create_device_data_table.php
@@ -15,7 +15,7 @@ class CreateDeviceDataTable extends Migration
     {
         Schema::create('device_data', function (Blueprint $table) {
             $table->id();
-            $table->integer('acceleration');
+            $table->float('acceleration');
             $table->float('carbon');
             $table->integer('inclination');
             $table->time('time');
